requirements.txt
requests
pandas
snowflake-connector-python[pandas]
# boto3 